==> Scripts/Script_2_modelagem.Rmd <==
---
title: "Visualizacao Averting Expenditures Covid"
author: "Claudiano Neto"
date: "`r format(Sys.time(), '%d %B, %Y')`"
geometry: margin = 1cm
output:
  html_document:
    theme: flatly
    highlight: tango
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE, warning=FALSE}
library(scales)
library(tidyverse)
library(magrittr)
library(utf8)
library(foreign)
library(haven)
library(stargazer)
library(margins)
library(datasets)
```

```{r}
covid_basedados_imputado <- read_dta("C:/Users/User/Documents/covid/dados_dta/covid_basedados_imputado.dta")

base_cov <- covid_basedados_imputado
```


```{r}
#Transformando variaveis "confirmed", "deaths" e "gasto_equip_defensivo" em log.

base_cov$confirmed = as.numeric(base_cov$confirmed)
base_cov$confirmed = ifelse(base_cov$confirmed == 0 | is.na(base_cov$confirmed), base_cov$confirmed, log(base_cov$confirmed))

base_cov$deaths = as.numeric(base_cov$deaths)
base_cov$deaths = ifelse(base_cov$deaths == 0 | is.na(base_cov$deaths), base_cov$deaths, log(base_cov$deaths))

base_cov$gasto_equip_defensivo = as.numeric(base_cov$gasto_equip_defensivo)
base_cov$gasto_equip_defensivo = ifelse(base_cov$gasto_equip_defensivo == 0 | is.na(base_cov$gasto_equip_defensivo), base_cov$gasto_equip_defensivo, log(base_cov$gasto_equip_defensivo))
```

```{r}
# dados <- base_cov_full %>% 
      # filter(gasto_equip_defensivo > 0) 

# lm(gasto_equip_defensivo ~  sexo + estcivil + morte_cov + contat_cov + filho_dep + idade + deaths + confirmed + d_raca + renda + d_escolaridade, data = dados)

# ols_dados <- lm(gasto_equip_defensivo ~  confirmed, data = dados)

# summary(ols_dados)
```


# Testando modelos 
```{r}

ols_gd <- lm(gasto_equip_defensivo ~  sexo + estcivil + morte_cov + contat_cov + filho_dep + idade + deaths + confirmed + d_raca + renda + d_escolaridade, data = base_cov)

summary(ols_gd)
# Depois de analisar os dados acima, optei por escolher o modelo "reg_ed_prob"

```

# Calculando os Efeitos marginais do PROBIT

```{r}
# https://github.com/leeper/margins

probit_1 <- glm(equip_defensiv ~ sexo + estcivil + morte_cov + contat_cov + filho_dep + idade + deaths + confirmed + d_raca + renda + d_escolaridade, data = base_cov, family = binomial(link = "probit"))

probit_2 <- glm(equip_defensiv ~ sexo + estcivil + morte_cov + contat_cov + filho_dep + idade + deaths + confirmed + renda + d_escolaridade, data = base_cov, family = binomial(link = "probit"))

probit_3 <- glm(equip_defensiv ~ sexo + estcivil + morte_cov + contat_cov + filho_dep + idade + deaths + confirmed + d_escolaridade, data = base_cov, family = binomial(link = "probit"))

probit_4 <- glm(equip_defensiv ~ sexo + estcivil + morte_cov + contat_cov + filho_dep + deaths + confirmed + d_escolaridade, data = base_cov, family = binomial(link = "probit"))

probit_5 <- glm(equip_defensiv ~ sexo + estcivil + morte_cov + contat_cov + idade + deaths + confirmed + d_escolaridade, data = base_cov, family = binomial(link = "probit"))

# Esse modelo ficou bastante interessante. Melhor discutir com cadu e Roberta
probit_6 <- glm(equip_defensiv ~ sexo + estcivil + morte_cov + contat_cov + deaths + confirmed + d_escolaridade, data = base_cov, family = binomial(link = "probit"))

stargazer(probit_1, probit_2, probit_3, probit_4, probit_5, probit_6, type = 'text')
# considerando os modelos GLM, escolho o modelo probit_6

```

```{r}
stargazer(ols_gd, probit_6, type = 'text')
```

```{r}
# valores marginais 

# https://github.com/leeper/margins

#Modelo OlS
summary(ols_gd)
plot(margins(ols_gd))
summary(margins(ols_gd))

#Modelo PROBIT
summary(probit_6)
plot(margins(probit_6))
summary(margins(probit_6))
```

```{r}
# Calculando intervalo de confianca

confint((ols_gd))
confint(probit_6)
```

```{r}
# Wald Test

library(aod)
wald.test(b = coef(ols_gd), Sigma = vcov(ols_gd), Terms = 6:10)
wald.test(b = coef(probit_6), Sigma = vcov(probit_6), Terms = 4:6)

```


```{r}
# McFadden's Pseudo R-squared
probit_60 <- update(probit_6, formula= equip_defensiv ~ 1)   # Calcula o LL mantendo Y constante.
McFadden<- 1-as.vector(logLik(probit_6)/logLik(probit_60))
McFadden


# acredito que o teste McFadden nao seja aplicav�l para ols de variavel dependente continua
ols_gd0 <- update(ols_gd, formula= gasto_equip_defensivo ~ 1)   # Calcula o LL mantendo Y constante.
McFadden_ols <- 1-as.vector(logLik(ols_gd)/logLik(ols_gd0))
McFadden_ols
# resultado de 0.12 está fora do intervalo sugerido por Louviere, e abaixo do obtido com a base_imp.

# o valor de 0.22 está dentro do range de 0.20 - 0.40 sugerido por Louviere et al (2000). "Stated Choice Methods: Analysis and Applications". Cambridge University Press., pagina 55.

# A explicacao para essa parte vem: https://portal.uni-freiburg.de/empiwifo/teaching/summer-term-09/materials-microeconometrics/probit_7-5-09.pdf 
```

```{r}
# A respeito do McFadden: Ele compara a probabilidade de log irrestrita Lur para o modelo que estamos estimando e a probabilidade de log restrita Lr com apenas uma interceptação. Se as variáveis independentes não tiverem poder explicativo, o modelo restrito será o mesmo que o modelo irrestrito e R ao quadrado será 0.
# Para ver mais detalhes sobre teste de hipoteses, ver: https://www.empiwifo.uni-freiburg.de/teaching/summer-term-09/materials-microeconometrics/probit_7-5-09.pdf .
# Para encontrar a diferença de desvio para os dois modelos (ou seja, a estat�?stica do teste), podemos calcular a alteração no desvio e testá-la usando um teste do qui quadrado - a mudança no desvio distribu�?da como qui quadrado na mudança nos graus de liberdade.
# Diggle, P.J., Liang, K.-Y., Zeger, S.L., 1994. Analysis of longitudinal data. Oxford, Clarendon Press, 253 p. Draper, N.R., Smith, H., 1998. Applied Regression Analysis. New York, John Wiley & Sons, Inc., 706 p.
```


```{r}
library(rsq)

# likelihood-ratio-based R^2
rsq(probit_6, adj = FALSE, type = 'lr')
rsq(probit_6, adj=TRUE)


#Calculate the likelihood-ratio-based R^2 for generalized linear models.
rsq(ols_gd, adj = FALSE, type = 'lr')
rsq.lr(ols_gd, adj = TRUE)
```

```{r}

## change in deviance
# with(ols_gd, null.deviance - deviance)
# ## change in degrees of freedom
# with(ols_gd, df.null - df.residual)
# ## chi square test p-value
# with(ols_gd, pchisq(null.deviance - deviance, df.null - df.residual, lower.tail = FALSE))
# # Razao de verossimilhanca
# logLik(ols_gd)


## change in deviance
with(probit_6, null.deviance - deviance)
## change in degrees of freedom
with(probit_6, df.null - df.residual)
## chi square test p-value
with(probit_6, pchisq(null.deviance - deviance, df.null - df.residual, lower.tail = FALSE))
# Razao de verossimilhanca
logLik(probit_6)
```

```{r}
# O qui-quadrado de 234.8032 com 11 graus de liberdade e um valor p associado < 0,001 nos diz que nosso modelo como um todo se encaixa significativamente melhor que um modelo vazio. As vezes, isso é chamado de teste da razão de verossimilhança (o desvio residual é -2 * log verossimilhança). A probabilidade de log do modelo foi -791.4241.
```


```{r}
# Multiple linear regression
# 
# summary(reg_ed_prob)
# confint(reg_ed_prob, level=0.95)   # intervalo de confianca
# anova(reg_ed_prob)    # teste anova
```

```{r}
# # Plotting regression line
 plot(log(gasto_equip_defensivo) ~ d_escolaridade + sexo + estcivil + morte_cov + contat_cov + filho_dep + idade + d_raca + renda, data = base_cov)
```

```{r}
# Plotando valores marginais

#https://cran.r-project.org/web/packages/sjPlot/vignettes/plot_marginal_effects.html

# library(sjPlot)
# library(ggeffects)
# library(ggplot2)
# theme_set(theme_sjplot())
# 
# plot_model(reg_ed_prob, type = "pred", terms = "idade")
# plot_model(reg_ed_prob, type = "pred", terms = "renda")
# plot_model(reg_ed_prob, type = "pred", terms = "filho_dep")
```

```{r, message=FALSE, warning=FALSE}
# confint(teste7)
```

```{r, message=FALSE, warning=FALSE}
#   https://stats.idre.ucla.edu/r/dae/probit-regression/
#   https://stats.idre.ucla.edu/stat/data/intro_r/intro_r_interactive.html#(61)
# require(aod)
# wald.test(b = coef(teste7), Sigma = vcov(teste7), Terms = 4:6)
```

```{r, message=FALSE, warning=FALSE}
#depois de montar a equação final 
#acessar https://jbhender.github.io/Stats506/F18/GP/Group14.html 
# para ver como fazer graficos interessantes

# https://stats.idre.ucla.edu/r/dae/ordinal-logistic-regression/

```

```{r}
# Pesquisar novamente o site https://www.statmethods.net/stats/regression.html
```

# Analisando Heckman
```{r}
library(sampleSelection)

base_cov$letal <- base_cov$deaths/base_cov$confirmed

heck_cov1 = heckit(
  selection = equip_defensiv ~ sexo + morte_cov + estcivil,
  outcome = gasto_equip_defensivo ~ sexo + morte_cov + contat_cov + filho_dep + idade + d_raca + renda + d_escolaridade,
  data = base_cov,
  method = "2step"
)
summary(heck_cov1)

heck_cov2 = heckit(
  selection = equip_defensiv ~ sexo + letal + estcivil,
  outcome = gasto_equip_defensivo ~ sexo + letal + contat_cov + filho_dep + idade + d_raca + renda + d_escolaridade,
  data = base_cov,
  method = "2step"
)
summary(heck_cov2)

heck_cov3 = heckit(
  selection = equip_defensiv ~ sexo + morte_cov + estcivil,
  outcome = gasto_equip_defensivo ~ sexo + letal + contat_cov + filho_dep + idade + d_raca + renda + d_escolaridade,
  data = base_cov,
  method = "2step"
)
summary(heck_cov3)


heck_cov4 = heckit(
  selection = equip_defensiv ~ sexo + morte_cov + estcivil,
  outcome = gasto_equip_defensivo ~ sexo + letal + filho_dep + idade + d_raca + renda + d_escolaridade,
  data = base_cov,
  method = "2step"
)
summary(heck_cov4)

heck_cov5 = heckit(
  selection = equip_defensiv ~ sexo + d_raca + estcivil + d_escolaridade,
  outcome = gasto_equip_defensivo ~ sexo + deaths + contat_cov + filho_dep + idade,
  data = base_cov,
  method = "2step"
)
summary(heck_cov5)

# O modelo 6 parece ter melhores resultados
heck_cov6 = heckit(
  selection = equip_defensiv ~ sexo + morte_cov + estcivil,
  outcome = gasto_equip_defensivo ~ sexo + deaths + contat_cov + filho_dep + idade + d_raca + renda + d_escolaridade,
  data = base_cov,
  method = "2step"
)
summary(heck_cov6)
```


```{r}
summary(lm(gasto_equip_defensivo ~ deaths, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ confirmed, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ sexo, data =  base_cov)) # NAO SIG
summary(lm(gasto_equip_defensivo ~ morte_cov, data =  base_cov)) 
summary(lm(gasto_equip_defensivo ~ contat_cov, data =  base_cov)) 
summary(lm(gasto_equip_defensivo ~ filho_dep, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ letal, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ renda, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ d_escolaridade, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ idade, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ d_raca, data =  base_cov))
summary(lm(gasto_equip_defensivo ~ estcivil, data =  base_cov))
```
